Fix test_get_float
