+++ docker-compose.yml
@@ -69,15 +69,12 @@ services:
     ports:
       - "8011:8011"
     environment:
-      SECRET_KEY: asdfb
-      MITX_ONLINE_ADMIN_EMAIL: admin@localhost
-      
-    #   << : *py-environment
-    #   PORT: 8011
-    #   # these are necessary to get the hostname from the varnish proxy
-    #   USE_X_FORWARDED_HOST: true
-    #   USE_X_FORWARDED_PORT: true
-    # env_file: .env
+      << : *py-environment
+      PORT: 8011
+      # these are necessary to get the hostname from the varnish proxy
+      USE_X_FORWARDED_HOST: true
+      USE_X_FORWARDED_PORT: true
+    env_file: .env
     links:
       - db
       - redis
