+++ scripts/test/python_tests.sh
@@ -20,6 +20,6 @@ function run_test {
 run_test ./scripts/test/detect_missing_migrations.sh
 run_test ./scripts/test/no_auto_migrations.sh
 run_test ./scripts/test/openapi_spec_check.sh
-run_test pytest -n logical -k test_get_float
+run_test pytest -n logical
 
 exit $status
